bump depth col from UInt16 -> UInt32

+++ athena/utils/constants.py
@@ -11,7 +11,7 @@ DATAFRAME_TYPES = {
     "region_length": pl.UInt16,
     "depth_bin_start": pl.UInt32,
     "depth_bin_end": pl.UInt32,
-    "depth": pl.UInt16,
+    "depth": pl.UInt32,
 }
 
 # value to use for normalising against for multi sample calculations
